requirements.txt
newspaper3k
boto3
shortuuid